МСХОАД/pythonProject1/requirement.txt
pandas
matplotlib
numpy
scipy
IPython
sklear
nmglearn
graphviz